broke prediction trajectory into a few files
